More rust cleanup
- Refactored BinaryReader and BinaryWriter
- Added some warnings to high_level_il and medium_level_il modules that they are unstable
- Add BinaryReader and BinaryWriter tests
- Changed BinaryView len to return u64 (that is what the core returns)
- Misc formatting changes
- Remove extern uses in lib.rs

+++ plugins/pdb-ng/src/lib.rs
@@ -552,7 +552,7 @@ impl CustomDebugInfoParser for PDBParser {
     ) -> bool {
         if is_pdb(debug_file) {
             match self.load_from_file(
-                &debug_file.read_vec(0, debug_file.len()),
+                &debug_file.read_vec(0, debug_file.len() as usize),
                 debug_info,
                 view,
                 &progress,
