+++ view/bintxt/src/srec.rs
@@ -212,7 +212,7 @@ impl CustomBinaryViewType for SRecViewConstructor {
         parent: &BinaryView,
         builder: CustomViewBuilder<'builder, Self>,
     ) -> Result<CustomView<'builder>, ()> {
-        let bytes = parent.len();
+        let bytes = parent.len() as usize;
         let mut buf = vec![0; bytes];
         let bytes_read = parent.read(&mut buf, 0);
         if bytes_read != bytes {
